move the dummy level / schema renderer classes to reasonable packages
